+++ src/Plugin.php
@@ -330,6 +330,7 @@ class Plugin extends BasePlugin
         $event->rules['simple-form/integrations/resend'] = 'simple-form/integrations/resend';
 
         $event->rules['simple-form/submissions'] = 'simple-form/submissions/index';
+        $event->rules['simple-form/submissions/export'] = 'simple-form/submissions/export';
         $event->rules['simple-form/submissions/<submissionId:\d+>'] = 'simple-form/submissions/view';
         $event->rules['simple-form/submissions/toggle-status'] = 'simple-form/submissions/toggle-status';
         $event->rules['simple-form/settings'] = 'simple-form/settings/index';